Solved a few issues in request.rs
Removed the Encrypt trait and its implementations as they were not used and not implemented. Will add them when I fully realize what's going on.
Implemented the TryFrom trait for Request to parse the HTTP request.
Added a FromStr implementation for Method to parse the method string.
Added a helper function get_next_word to assist in parsing the request.
Made Request derive Debug for easier printing.